requirements.txt
# Dashboard
shiny>=1.0.0

# Pipeline - HTTP
httpx>=0.27.0

# Pipeline - Parsing
feedparser>=6.0.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Pipeline - Enrichment
google-genai>=1.0.0

# Pipeline - Notifications
resend>=2.0.0

# Pipeline - Deduplication
rapidfuzz>=3.0.0
url-normalize>=1.4.0

# Pipeline - Models and config
pydantic>=2.0.0
pyyaml>=6.0.0

# Pipeline - Templates
jinja2>=3.1.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
